showing movie and reviews of that movie in movieview. pending fixing some issues like asynchronous code and lifecycles and the refreshtoken method.

+++ src/services/authService.ts
@@ -129,7 +129,7 @@ export class AuthService {
     return returnData;
   }
 
-  //REFRESH TOKEN---------------------------------------------------
+  //REFRESH TOKEN---------------------------------------------------PENDING FIXING REFRESHTOKEN
   static async refreshToken(): Promise<IReturnData> {
     const returnData: IReturnData = {
       error: false,
@@ -251,17 +251,17 @@ export class AuthService {
       return { Authorization: `ERROR` };
     }
 
-    //const minutes = AuthService.checkMinutesOfTokensLife(payload.exp);
+    const minutes = AuthService.checkMinutesOfTokensLife(payload.exp);
 
     try {
-      //if (minutes >= 10) {
+      if (minutes >= 10) {
         const result = await AuthService.refreshToken();
         if (result.error) {
           AuthService.removeTokensAndClearStore();
           return { Authorization: `ERROR` };
         }
         accessToken = localStorage.getItem(Tokens.ACCESS_TOKEN);
-      //}
+      }
     } catch (err) {
       AuthService.removeTokensAndClearStore();
       return { Authorization: `ERROR` };
